bump go; streaming now requires streamkey in db
go bump to 1.22 for url path parameters
Streamer struct gets passed around, which contains name of the current
streamer, their authtoken and the streamkey.
this gets created and set on stream start, and is then stored in the
streamMap which still uses stream key as the key. this struct is only
used to know the streamer name for now to return in /api/status
/api/status now requires a stream key in the path, and should reject
anything that doesn't exist in the db. it gets the valid stream keys
when the url is called, so you can dynamically add more valid stream
keys while the server is running.

also *dbPool is now passed around as a parameter instead of making an
instance of a connection or similar.
seems solid (knock on wood), not sure what happens if multiple people are streaming and
one tries to fetch /api/status/{streamkey}